Create cal.md
It might not be obvious to the EICUG what we intend to do with the calendar. In the task force, we thought that it would be a good service for the community to provide a calendar of the EICUG meetings that everyone could imported in / synchronized with their calendar. 

Thus, I have created this page. If you find this helpful, we should replace the current link in the main menu with a link to this calendar page.